Rename more IR imports
