+++ ffc/ir/uflacs/analysis/factorization.py
@@ -9,9 +9,9 @@
 import logging
 from functools import singledispatch
 
-from ffc.uflacs.analysis.graph import ExpressionGraph
-from ffc.uflacs.analysis.modified_terminals import (analyse_modified_terminal,
-                                                    strip_modified_terminal)
+from ffc.ir.uflacs.analysis.graph import ExpressionGraph
+from ffc.ir.uflacs.analysis.modified_terminals import (analyse_modified_terminal,
+                                                       strip_modified_terminal)
 from ufl import as_ufl, conditional
 from ufl.classes import Argument, Conditional, Division, Product, Sum, Zero, Conj
