+++ ffc/ir/uflacs/integralgenerator.py
@@ -12,7 +12,7 @@ import logging
 
 import ufl
 from ffc.codegeneration.C.cnodes import pad_dim, pad_innermost_dim
-from ffc.uflacs.elementtables import piecewise_ttypes
+from ffc.ir.uflacs.elementtables import piecewise_ttypes
 
 
 logger = logging.getLogger(__name__)
